% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheby.R
\name{cheb}
\alias{cheb}
\title{cheby}
\usage{
cheb(x)
}
\arguments{
\item{x}{= uses data given to solve for the percentage of data within a standard deviation}
}
\value{
a numerical value stating the amount of data present within a certain number of standard deviation
}
\description{
cheby
}
