% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeo.R
\name{hypergeo}
\alias{hypergeo}
\title{Hypergeometric Function}
\usage{
hypergeo(iter = 100, N = 20, r = 12, n = 6)
}
\arguments{
\item{iter}{= iterations}

\item{N}{= total size of group}

\item{r}{= specific sub group}

\item{n}{= sample size}
}
\value{
a histogram plot
}
\description{
This function will make it easier to plot one sample using the hypergeomtric distribution
}
