% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Given a sample, calculate the point estimate and a (1-alpha)*100 percent
ci for a function over a given amount of iterations.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations to use for calculation}

\item{x}{the sample to be used}

\item{fun}{the function to be applied onto our sample}

\item{alpha}{used to determine the (1-alpha)*100 percent ci}

\item{cx}{the relative scaling of the text/symbols for plots}

\item{...}{more parameters for plotting histogram}
}
\value{
a list, containing the data for the function, the ci and the provided sample
}
\description{
Given a sample, calculate the point estimate and a (1-alpha)*100 percent
ci for a function over a given amount of iterations.
}
