% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCLT.R
\name{mycltMean}
\alias{mycltMean}
\title{CLT Mean Function}
\usage{
mycltMean(n, iter)
}
\arguments{
\item{n}{the size of the random sample}

\item{iter}{the number of iterations within the simulation}
}
\value{
the vector containing the vector of mean values for each iteration's
random sample
}
\description{
This will generate an n-sized random sample from the uniform distribution for a given
amount of iterations. After doing this, it will calculate the mean and then create a histogram for that sample.
}
