% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve Functioni}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{= mean}

\item{sigma}{= standard deviation}

\item{a}{= bounds}
}
\value{
a curve with a shaded area
}
\description{
This function will make it easier to plot a shaded area within a curve by calculating the area for you given the certain parameters.
}
