% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{ntickets}
\alias{ntickets}
\title{Ntickets Function}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats, default 200}

\item{gamma}{Probability of passenger not showing up, default 0.02}

\item{p}{Probability that N or fewer passengers do show up, default 0.95}
}
\value{
Produces a graph for the binomial and normal distribution for objective v n Also prints a named list of nd, nc, N, p, and gamma.
}
\description{
Ntickets Function
}
